Handle both *_gnu and *_gnullvm import libraries from the same command

+++ crates/tools/gnu/src/main.rs
@@ -3,14 +3,7 @@ use std::io::prelude::*;
 use std::process::{Command, Stdio};
 
 fn main() {
-    for cmd in ["dlltool", "ar", "objcopy"] {
-        if Command::new(cmd).stdout(Stdio::null()).stderr(Stdio::null()).spawn().is_err() {
-            eprintln!("Could not find {}. Is it in your $PATH?", cmd);
-            return;
-        }
-     }
-
-    const ALL_PLATFORMS: [&str; 2] = ["x86_64_gnu", "i686_gnu"];
+    const ALL_PLATFORMS: [&str; 4] = ["x86_64_gnu", "i686_gnu", "x86_64_gnullvm", "aarch64_gnullvm"];
     let mut platforms = BTreeSet::new();
     for platform in std::env::args().skip(1) {
         if ALL_PLATFORMS.contains(&&*platform) {
@@ -28,11 +21,19 @@ fn main() {
     };
 
     for platform in platforms {
-        build_platform(&platform);
+        let tools = if platform.ends_with("_gnu") { &["dlltool", "ar", "objcopy"][..] } else { &["llvm-dlltool", "llvm-ar"][..] };
+        for tool in tools {
+            if Command::new(tool).stdout(Stdio::null()).stderr(Stdio::null()).spawn().is_err() {
+                eprintln!("Could not find {}. Is it in your $PATH?", tool);
+                return;
+            }
+        }
+
+        build_platform(&platform, tools[0], tools[1]);
     }
 }
 
-fn build_platform(platform: &str) {
+fn build_platform(platform: &str, dlltool: &str, ar: &str) {
     println!("Platform: {}", platform);
 
     let libraries = lib::libraries();
@@ -41,17 +42,17 @@ fn build_platform(platform: &str) {
     std::fs::create_dir_all(&output).unwrap();
 
     for (library, functions) in &libraries {
-        build_library(&output, library, functions, platform);
+        build_library(&output, dlltool, library, functions, platform);
     }
 
-    build_mri(&output, &libraries);
+    build_mri(&output, ar, &libraries);
 
     for library in libraries.keys() {
         std::fs::remove_file(output.join(format!("lib{}.a", library))).unwrap();
     }
 }
 
-fn build_library(output: &std::path::Path, library: &str, functions: &BTreeMap<String, lib::CallingConvention>, platform: &str) {
+fn build_library(output: &std::path::Path, dlltool: &str, library: &str, functions: &BTreeMap<String, lib::CallingConvention>, platform: &str) {
     println!("{}", library);
 
     // Note that we don't use set_extension as it confuses PathBuf when the library name includes a period.
@@ -80,7 +81,7 @@ EXPORTS
 
     drop(def);
 
-    let mut cmd = Command::new("dlltool");
+    let mut cmd = Command::new(dlltool);
     cmd.current_dir(output);
 
     if platform.eq("i686_gnu") {
@@ -94,50 +95,55 @@ EXPORTS
     cmd.arg("-m");
     cmd.arg(match platform {
         "i686_gnu" => "i386",
-        "x86_64_gnu" => "i386:x86-64",
+        "x86_64_gnu" | "x86_64_gnullvm" => "i386:x86-64",
+        "aarch64_gnullvm" => "arm64",
         _ => unreachable!(),
     });
-    // Work around https://sourceware.org/bugzilla/show_bug.cgi?id=29497
-    cmd.arg("-f");
-    cmd.arg(match platform {
-        "i686_gnu" => "--32",
-        "x86_64_gnu" => "--64",
-        _ => unreachable!(),
-    });
-    // Ensure consistency in the prefixes used by dlltool.
-    cmd.arg("-t");
-    if library.contains('.') {
-        cmd.arg(format!("{}_", library).replace('.', "_").replace('-', "_"));
-    } else {
-        cmd.arg(format!("{}_dll_", library).replace('-', "_"));
+    if dlltool == "dlltool" {
+        // Work around https://sourceware.org/bugzilla/show_bug.cgi?id=29497
+        cmd.arg("-f");
+        cmd.arg(match platform {
+            "i686_gnu" => "--32",
+            "x86_64_gnu" => "--64",
+            _ => unreachable!(),
+        });
+        // Ensure consistency in the prefixes used by dlltool.
+        cmd.arg("-t");
+        if library.contains('.') {
+            cmd.arg(format!("{}_", library).replace('.', "_").replace('-', "_"));
+        } else {
+            cmd.arg(format!("{}_dll_", library).replace('-', "_"));
+        }
     }
     cmd.output().unwrap();
 
-    // Work around lack of determinism in dlltool output, and at the same time remove
-    // unnecessary sections and symbols.
-    std::fs::rename(output.join(format!("lib{}.a", library)), output.join("tmp.a")).unwrap();
-    let mut cmd = Command::new("objcopy");
-    cmd.current_dir(output);
-    cmd.arg("--remove-section=.bss");
-    cmd.arg("--remove-section=.data");
-    cmd.arg("--strip-unneeded-symbol=fthunk");
-    cmd.arg("--strip-unneeded-symbol=hname");
-    cmd.arg("--strip-unneeded-symbol=.file");
-    cmd.arg("--strip-unneeded-symbol=.text");
-    cmd.arg("--strip-unneeded-symbol=.data");
-    cmd.arg("--strip-unneeded-symbol=.bss");
-    cmd.arg("--strip-unneeded-symbol=.idata$7");
-    cmd.arg("--strip-unneeded-symbol=.idata$5");
-    cmd.arg("--strip-unneeded-symbol=.idata$4");
-    cmd.arg("tmp.a");
-    cmd.arg(format!("lib{}.a", library));
-    cmd.output().unwrap();
+    if dlltool == "dlltool" {
+        // Work around lack of determinism in dlltool output, and at the same time remove
+        // unnecessary sections and symbols.
+        std::fs::rename(output.join(format!("lib{}.a", library)), output.join("tmp.a")).unwrap();
+        let mut cmd = Command::new("objcopy");
+        cmd.current_dir(output);
+        cmd.arg("--remove-section=.bss");
+        cmd.arg("--remove-section=.data");
+        cmd.arg("--strip-unneeded-symbol=fthunk");
+        cmd.arg("--strip-unneeded-symbol=hname");
+        cmd.arg("--strip-unneeded-symbol=.file");
+        cmd.arg("--strip-unneeded-symbol=.text");
+        cmd.arg("--strip-unneeded-symbol=.data");
+        cmd.arg("--strip-unneeded-symbol=.bss");
+        cmd.arg("--strip-unneeded-symbol=.idata$7");
+        cmd.arg("--strip-unneeded-symbol=.idata$5");
+        cmd.arg("--strip-unneeded-symbol=.idata$4");
+        cmd.arg("tmp.a");
+        cmd.arg(format!("lib{}.a", library));
+        cmd.output().unwrap();
 
-    std::fs::remove_file(output.join("tmp.a")).unwrap();
+        std::fs::remove_file(output.join("tmp.a")).unwrap();
+    }
     std::fs::remove_file(output.join(format!("{}.def", library))).unwrap();
 }
 
-fn build_mri(output: &std::path::Path, libraries: &BTreeMap<String, BTreeMap<String, lib::CallingConvention>>) {
+fn build_mri(output: &std::path::Path, ar: &str, libraries: &BTreeMap<String, BTreeMap<String, lib::CallingConvention>>) {
     let mri_path = output.join("unified.mri");
     let mut mri = std::fs::File::create(&mri_path).unwrap();
     println!("Generating {}", mri_path.to_string_lossy());
@@ -150,7 +156,7 @@ fn build_mri(output: &std::path::Path, libraries: &BTreeMap<String, BTreeMap<Str
 
     mri.write_all(b"SAVE\nEND\n").unwrap();
 
-    let mut cmd = Command::new("ar");
+    let mut cmd = Command::new(ar);
     cmd.current_dir(output);
     cmd.arg("-M");
     cmd.stdin(std::fs::File::open(&mri_path).unwrap());
